Added new resourcea
Added 2 new resources
Sound settings, this holds all the information settings that a theme can have, it is also what we will use to calc what theme to give to the user.
Sound setting map: this will be a single resource that holds the theme and it's settings into a singular resource dictonary so that we do not have to search for it later, we can simply load this dict  get on key and retrieve the correct information.